requirements.txt
streamlit
streamlit-searchbox
pillow
unidecode
pandas
numpy
altair
streamlit-screen-stats
streamlit-browser-session-storage
streamlit-local-storage